streamlit_requirements.txt
# Streamlit
streamlit==1.29.0

# HTTP Client
requests==2.31.0

# Image Processing
pillow==10.1.0

# Utilities
python-dotenv==1.0.0

